// src/app/models/predio.ts
export interface Predio {
  item: number;
  codigoNuevo: string;
  nombreApellido: string;
  dniRuc: string;
  estadoCivil: string;
  sector: string;
  distrito: string;
  provincia: string;
  departamento: string;
  unidadCatastral: string;
  progesivaInicial: string;
  progresivaFinal: string;
  lado: string;
  usoDelPredio: string;
  tipoAfectacion: string;
  tipoPredio: string;
  zonificacion: string;
  areaGrafica: number;
  areaDocumentoM2: number;
  areaDocumentoHa: number;
  areaTotalM2: number;
  areaTotalHa: number;
  areaAfectadaM2: number;
  areaAfectadaHa: number;
  areaRemanenteM2: number;
  areaRemanenteHa: number;
  areaDirectaM2: number;
  areaIndirectaM2: number;
  modulo: string;
  nNiveles: number;
  areaTechadaDirectaM2: number;
  areaTechadaIndirectaM2: number;
  areaTechadaTotalM2: number;
  nombreCientifico: string;
  nombreComun: string;
  variedad: string;
  edad: number;
  unidadMedida: string;
  altura: number;
  diametro: number;
  numeroPlantas: number;
  areaCultivosM2: number;
  costo: number;
  descripcion: string;
  descripcionDetallada: string;
  alturaObra: number;
  longitudObra: number;
  anchoEspesor: number;
  areaObra: number;
  unidadObra: string;
  ubicacionObra: string;
  condicionJuridica: string;
  estadoPredio: string;
  partidaElectronica: string;
  documentoAdjunto: string;
  afectacionPor: string;
  antecedentesPropiedad: string;
  cargasGravamenes: string;
  duplicidadPartidas: string;
  naturalezaTitular: string;
  nombrePredio: string;
  clasificacionDL1192: string;
  observaciones: string;
  conclusiones: string;
  recomendaciones: string;
}